requirements.txt
streamlit
openai
python-dotenv
pandas
openpyxl
PyMuPDF
python-docx
fpdf2
requests
joblib
numpy
scikit-learn
sentence-transformers
langdetect
supabase==1.2.0
bcrypt==4.1.3
plotly==5.22.0
cryptography

